moved sign in handling to sessions folder. And implemented sessions
