+++ instagram_web/__init__.py
@@ -1,6 +1,7 @@
 from app import app
 from flask import render_template
 from instagram_web.blueprints.users.views import users_blueprint
+from instagram_web.blueprints.sessions.views import sessions_blueprint
 from flask_assets import Environment, Bundle
 from .util.assets import bundles
 
@@ -8,6 +9,8 @@ assets = Environment(app)
 assets.register(bundles)
 
 app.register_blueprint(users_blueprint, url_prefix="/users")
+app.register_blueprint(sessions_blueprint, url_prefix="/sessions")
+
 
 @app.errorhandler(500)
 def internal_server_error(e):
