+++ instagram_web/blueprints/sessions/templates/sessions/new.html
@@ -8,7 +8,7 @@ Sign In
 {% endblock %}
 
 {% block content %}
-<form action='{{url_for("users.signed_in")}}' method='POST'>
+<form action='{{url_for("sessions.sign_in")}}' method='POST'>
     <input type="hidden" name="csrf_token" value="{{ csrf_token() }}" />
     <div class="form-group">
         <label for="exampleInputU">Username</label>
